requirements.txt
streamlit
PyPDF2
pdfplumber
python-docx
pandas
spacy
python-magic-bin